Add the first protoype of RNN
This notebook has an RNN w/ LSTMs trained on
Mahindra & Mahindra stocks. It has the complete
code from dataset cleaning to model prediction.
It will serve as the backbone to create the
complete system capable of handling multiple
datasets at once and its model will be utilised
for transfer learning.